fix : 사이드바 합치기 및 누르면 해당 라우터로 가지는 기능 추가
