+++ src/layout/Layout.tsx
@@ -1,17 +1,15 @@
 import React from 'react';
-import SidebarIsLogged from './SidebarIsLogged';
-import { useAuthStore } from '../stores/authStore';
-import SidebarIsNotLogged from './SidebarIsNotLogged';
 import { Outlet } from 'react-router-dom';
 import Header from './Header';
+import Sidebar from './Sidebar';
 
 export default function Layout({ children }: { children?: React.ReactNode }) {
-    const isAuthenticated = useAuthStore((state) => state.isLoggedIn);
     return (
         <div>
             <Header />
             <div className="flex flex-1 min-h-screen">
-                {isAuthenticated ? <SidebarIsLogged /> : <SidebarIsNotLogged />}
+                {/* {isAuthenticated ? <SidebarIsLogged /> : <SidebarIsNotLogged />} */}
+                <Sidebar />
                 <main className="flex-1 overflow-y-auto p-4">{children || <Outlet />}</main>
             </div>
         </div>
