+++ src/layout/Sidebar.tsx
@@ -1,0 +1,176 @@
+import { MenuItem } from '../types/Sidebar';
+import newPostIcon from '../assets/icons/newPostIcon.svg';
+import mate from '../assets/icons/homeIcon.svg';
+import search from '../assets/icons/Search.svg';
+import letter from '../assets/icons/letterIcon.svg';
+import notice from '../assets/icons/notificationIcon.svg';
+import history from '../assets/icons/profileIcon.svg';
+import settings from '../assets/icons/setting.svg';
+import { useState } from 'react';
+import SearchPanel from '../components/SearchPanel';
+import NoticePanel from '../components/NoticePanel';
+import { useNavigate } from 'react-router-dom';
+import { useAuthStore } from '../stores/authStore';
+
+export default function Sidebar(): React.ReactElement {
+    const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
+    const navigate = useNavigate();
+
+    const [isSearchMode, setIsSearchMode] = useState<boolean>(false);
+    const [highlightedItemId, setHighlightedItemId] = useState<string | null>('mate');
+
+    const handleItemClick = (item: MenuItem, event: React.MouseEvent) => {
+        event.preventDefault();
+
+        if (item.isSearchTrigger) {
+            const nextSearchMode = !isSearchMode;
+            setIsSearchMode(nextSearchMode);
+            setHighlightedItemId(nextSearchMode ? item.id : null);
+            return;
+        }
+
+        // 클릭 시 색상 활성화
+        setIsSearchMode(false);
+        setHighlightedItemId(item.id === highlightedItemId ? null : item.id);
+
+        // 메뉴별 네비게이션 라우팅
+        switch (item.id) {
+            case 'share':
+                navigate('/write');
+                break;
+            case 'mate':
+                navigate('/channel');
+                break;
+            case 'history':
+                navigate('/myprofile');
+                break;
+            case 'settings':
+                navigate('/profileedit');
+                break;
+            // 편지/알림은 패널 열기 전용이므로 라우팅 없음
+        }
+    };
+
+    const handleCloseSearch = () => {
+        setIsSearchMode(false);
+        setHighlightedItemId(null);
+    };
+
+    const handleClosePanel = () => {
+        setHighlightedItemId(null);
+    };
+
+    const isExpanded = highlightedItemId !== null;
+    const shouldShrinkSidebar = highlightedItemId === 'search' || highlightedItemId === 'notice';
+    const menuPanelWidth = shouldShrinkSidebar ? 'w-[70px]' : 'w-[235px]';
+
+    //로그인 여부에 따라 메뉴 설정
+    const menuItems: MenuItem[] = isLoggedIn
+        ? [
+              {
+                  id: 'share',
+                  icon: <img src={newPostIcon} />,
+                  text: '여행 나누기',
+                  path: '/share',
+              },
+              { id: 'mate', icon: <img src={mate} />, text: '여행 이야기', path: '/mate' },
+              {
+                  id: 'search',
+                  icon: <img src={search} />,
+                  text: '검색',
+                  path: '/search',
+                  isSearchTrigger: true,
+              },
+              { id: 'letter', icon: <img src={letter} />, text: '편지', path: '/letter' },
+              {
+                  id: 'notice',
+                  icon: <img src={notice} />,
+                  text: '여정 알림판',
+                  path: '/notice',
+              },
+              {
+                  id: 'history',
+                  icon: <img src={history} />,
+                  text: '나의 여정 기록',
+                  path: '/history',
+              },
+              {
+                  id: 'settings',
+                  icon: <img src={settings} />,
+                  text: '설정',
+                  path: '/settings',
+              },
+          ]
+        : [
+              {
+                  id: 'mate',
+                  icon: <img src={mate} />,
+                  text: '여행 이야기',
+                  path: '/mate',
+              },
+              {
+                  id: 'search',
+                  icon: <img src={search} />,
+                  text: '검색',
+                  path: '/search',
+                  isSearchTrigger: true,
+              },
+          ];
+
+    return (
+        <nav className="min-h-screen bg-white flex transition-all duration-300">
+            <div
+                className={`p-4 overflow-y-auto bg-[var(--color-sideBody)] transition-all duration-300 ease-in-out ${menuPanelWidth}`}
+            >
+                <ul className="items-start justify-start text-left mt-1.5">
+                    {menuItems.map((item) => {
+                        const isItemActive = item.id === highlightedItemId;
+                        const iconContainerClasses = `p-2 mb-2 rounded-md cursor-pointer transition duration-150 ease-in-out group flex items-center justify-center ${
+                            isItemActive ? 'bg-[var(--color-selected)]' : 'bg-[var(--color-sideBody)]'
+                        }`;
+                        return (
+                            <li
+                                key={item.id}
+                                className={`${iconContainerClasses.trim()} items-start justify-start text-left`}
+                                onClick={(e) => handleItemClick(item, e)}
+                                title={item.text}
+                            >
+                                <span
+                                    className={`w-5 h-5 block mr-3 flex-shrink-0 ${
+                                        isItemActive ? 'filter brightness-0 invert' : ''
+                                    }`}
+                                >
+                                    {item.icon}
+                                </span>
+                                <span
+                                    className="text-sm"
+                                    style={{
+                                        color: isItemActive ? 'white' : 'var(--color-notSelected)',
+                                        display: shouldShrinkSidebar ? 'none' : 'inline',
+                                    }}
+                                >
+                                    {item.text}
+                                </span>
+                            </li>
+                        );
+                    })}
+                </ul>
+            </div>
+
+            <div className="flex-grow overflow-hidden">
+                {isExpanded && (
+                    <div className="h-full w-full bg-white">
+                        {isSearchMode ? (
+                            <SearchPanel onClose={handleCloseSearch} />
+                        ) : (
+                            <>
+                                {highlightedItemId === 'notice' && <NoticePanel onClose={handleClosePanel} />}
+                                {highlightedItemId && highlightedItemId !== 'notice' && highlightedItemId !== 'search'}
+                            </>
+                        )}
+                    </div>
+                )}
+            </div>
+        </nav>
+    );
+}
